Feat : 보스 체력바 연동

+++ Assets/2.Private/KimSW/Scripts/StatusBarUI/EnemyStatusPanel.cs
@@ -21,7 +21,7 @@ public class EnemyStatusPanel : MonoBehaviour
         }
     }
 
-    public void SetEnemyHP(int hp)
+    public void SetEnemyHP(float hp)
     {
         hpView.SetEnemyHp(hp);
